# [Enhancement] optimize JNI call for JDBC external table