--- a/be/src/exec/vectorized/jdbc_scanner.h
+++ b/be/src/exec/vectorized/jdbc_scanner.h
@@ -4,11 +4,14 @@
 
 #include "column/chunk.h"
 #include "column/vectorized_fwd.h"
+#include "common/object_pool.h"
 #include "common/status.h"
 #include "exec/olap_common.h"
 #include "exec/pipeline/scan/scan_operator.h"
+#include "exprs/expr_context.h"
 #include "jni.h"
 #include "runtime/descriptors.h"
+#include "runtime/primitive_type.h"
 #include "runtime/runtime_state.h"
 #include "udf/java/java_udf.h"
 
@@ -50,7 +53,7 @@ public:
 private:
     void _init_profile();
 
-    Status _precheck_data_type(const std::string& java_class, SlotDescriptor* slot_desc);
+    StatusOr<PrimitiveType> _precheck_data_type(const std::string& java_class, SlotDescriptor* slot_desc);
 
     Status _init_jdbc_bridge();
 
@@ -64,30 +67,11 @@ private:
 
     Status _has_next(bool* result);
 
-    Status _get_next_chunk(jobject* chunk);
+    Status _get_next_chunk(jobject* chunk, size_t* num_rows);
 
-    Status _fill_chunk(jobject jchunk, ChunkPtr* chunk);
+    Status _fill_chunk(jobject jchunk, size_t num_rows, ChunkPtr* chunk);
 
     Status _close_jdbc_scanner();
-
-    template <PrimitiveType type, typename CppType>
-    void _append_data(Column* column, CppType& value);
-
-    template <typename CppType>
-    Status _append_value_from_result(jobject jval, std::function<CppType(jobject)> get_value_func,
-                                     SlotDescriptor* slot_desc, Column* column);
-
-    Status _append_datetime_val(JNIEnv* env, jobject jval, SlotDescriptor* slot_desc, Column* column);
-
-    Status _append_localdatetime_val(JNIEnv* env, jobject jval, SlotDescriptor* slot_desc, Column* column);
-
-    Status _append_date_val(JNIEnv* env, jobject jval, SlotDescriptor* slot_desc, Column* column);
-
-    Status _append_decimal_val(JNIEnv* env, jobject jval, SlotDescriptor* slot_desc, Column* column);
-
-    std::string _get_date_string(JNIEnv* env, jobject jval);
-
-    std::string _get_localdatetime_string(JNIEnv* env, jobject jval);
 
     JDBCScanContext _scan_ctx;
     // result tuple desc
@@ -95,7 +79,10 @@ private:
     // result column slot desc
     std::vector<SlotDescriptor*> _slot_descs;
     // java class name for each result column
-    std::vector<std::string> _column_class_name;
+    std::vector<std::string> _column_class_names;
+    std::vector<PrimitiveType> _result_column_types;
+    std::vector<ExprContext*> _cast_exprs;
+    ChunkPtr _result_chunk;
 
     std::unique_ptr<JVMClass> _jdbc_bridge_cls;
     std::unique_ptr<JVMClass> _jdbc_scanner_cls;
@@ -103,6 +90,7 @@ private:
 
     jmethodID _scanner_has_next;
     jmethodID _scanner_get_next_chunk;
+    jmethodID _scanner_result_rows;
     jmethodID _scanner_close;
     // JDBCUtil method
     jmethodID _util_format_date;
@@ -116,6 +104,8 @@ private:
 
     RuntimeProfile* _runtime_profile = nullptr;
     JDBCScannerProfile _profile;
+
+    ObjectPool _pool;
 
     static constexpr const char* JDBC_BRIDGE_CLASS_NAME = "com/starrocks/jdbcbridge/JDBCBridge";
     static constexpr const char* JDBC_SCAN_CONTEXT_CLASS_NAME = "com/starrocks/jdbcbridge/JDBCScanContext";