--- a/be/src/exprs/vectorized/cast_expr.h
+++ b/be/src/exprs/vectorized/cast_expr.h
@@ -15,9 +15,10 @@ namespace vectorized {
 
 class VectorizedCastExprFactory {
 public:
-    static Expr* from_thrift(const TExprNode& node);
+    static Expr* from_thrift(const TExprNode& node, bool exception_if_failed = false);
 
-    static Expr* from_type(const TypeDescriptor& from, const TypeDescriptor& to, Expr* child, ObjectPool* pool);
+    static Expr* from_type(const TypeDescriptor& from, const TypeDescriptor& to, Expr* child, ObjectPool* pool,
+                           bool exception_if_failed = false);
 };
 
 // cast Array to Array.