## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
for #9748 

## Problem Summary(Required) ：

optimize JNI call for JDBC external table:

```
mysql> select count(id_tinyint) from jdbc_tbl;
baseline: 0.16 sec patched: 0.10 sec
mysql> select count(id_tinyint),count(id_smallint),count(id_int),count(id_bigint),count(id_date), count(id_varchar),count(id_datetime),count(id_decimal) from jdbc_tbl;
baseline: 1.07 sec patched:0.50 sec
```


## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] I have added user document for my new feature or new function