@stdpain 2022-08-15
run starrocks_admit_test