@stdpain 2022-08-19
run starrocks_admit_test