@stdpain 2022-08-20
run starrocks_admit_test