@stdpain 2022-08-20
run all