@stdpain 2022-08-20
run starrocks_be_unittest